Separate out adapt code
