+++ animate/__init__.py
@@ -1,1 +1,2 @@
+from animate.adapt import *  # noqa
 from animate.metric import *  # noqa
